% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseq.R
\name{aggregateRiboSeq}
\alias{aggregateRiboSeq}
\title{aggregateRiboSeq}
\usage{
aggregateRiboSeq(x, genome, str)
}
\arguments{
\item{x}{data frame returned by \code{overlapsRiboSeq} function}

\item{genome}{\code{BSgenome} object containing reference genome}

\item{str}{character: +/-/* indicating forward, reverse or both DNA strands; required in case RiboSeq data
is frame agnostic}
}
\value{
Data frame with aggregated data
}
\description{
This function analyses the data frame returned by \code{overlapsRiboSeq} function and returns
modified data frame where each position (first:width) of selected Genomic Ranges has its own column.
This function attaches also sequences in its own column and prepares data for further analysis using
machine learning algorithms.
}
