% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadGWIPSdata}
\alias{loadGWIPSdata}
\title{loadGWIPSdata}
\usage{
loadGWIPSdata(forward, reverse)
}
\arguments{
\item{forward}{path to the BigWig file containing Riboseq data for forward strand}

\item{reverse}{path to the BigWig file containing Riboseq data for forward strand}
}
\description{
Loads data from GWIPS database (in BigWig format) for forward and reverse strands into global GRanges objects. As a result, global objects "gwips_forw" and "gwips_rev" become available for further analyses.
}
