% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseq.R
\name{overlapsRiboSeq}
\alias{overlapsRiboSeq}
\title{overlapsRiboSeq}
\usage{
overlapsRiboSeq(selected, riboseq, str)
}
\arguments{
\item{selected}{\code{GRanges} object containing genomic features selected by selectFeaturesGR function}

\item{riboseq}{\code{GRanges} object (gwips_forw or gwips_rev) containing RiboSeq coverage data}

\item{str}{character: +/-/* indicating forward, reverse or both DNA strands; required in case RiboSeq data is frame agnostic}
}
\value{
Data frame with all obervations
}
\description{
Returns data frame with witn overlapping regions between selected features and RiboSeq data.
}
\examples{
overlapsRiboSeq(selectec=ecoli_genes_granges, riboseq=gwips_forw, strand="+")

}
