% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseq.R
\name{randomGenomicRanges}
\alias{randomGenomicRanges}
\title{randomGenomicRanges}
\usage{
randomGenomicRanges(x, genome)
}
\arguments{
\item{x}{vector of widths}

\item{genome}{\code{BSGenome} object}
}
\value{
GenomicRange object containing randomly selected, non-overlaping ranges of widths requested by vector \code{x}
}
\description{
Generate random Genomic Ranges given \code{BSGenome} object. The resulting
\code{GRanges} object will have unique starts, but ranges (depending on the vector of lengths)
might have overlaps. The only check implemented is that the sum of widths in the supplied vector
should not exceed the length of \code{BSGenome} object.
}
\examples{
randomGenomicRanges(GRfromGFF, type="end", width=25, feature="exon")

}
