% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseq.R
\name{selectFeaturesGR}
\alias{selectFeaturesGR}
\title{selectFeaturesGR}
\usage{
selectFeaturesGR(GR, type = "start", width = 10, feature = "gene",
  randomranges = NULL)
}
\arguments{
\item{GR}{\code{GRanges} object - it is assumed that it is full genomic annotation imported from GFF}

\item{type}{Could be "start", "end" or "random"}

\item{width}{Integer indicating length of selected ranges}

\item{feature}{"gene" or "exon", to be selected from GFF annotation. "gene" is the default.}

\item{randomranges}{optional argumument, required if \code{type="random"} is supplied; this is a
vector of widths}
}
\value{
GenomicRange object containing only selected features
}
\description{
Returns new GRanges object based on the selection criteria.
}
\examples{
selectFeaturesGR(GRfromGFF, type="end", width=25, feature="exon")

}
