% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges.R
\name{randomGenomicRanges}
\alias{randomGenomicRanges}
\title{randomGenomicRanges}
\usage{
randomGenomicRanges(x, gr)
}
\arguments{
\item{x}{vector of widths}

\item{gr}{\code{GRanges} object}
}
\value{
GenomicRange object containing randomly selected uniquue ranges of widths requested by vector \code{x}
}
\description{
Generate random Genomic Ranges given \code{GRanges} object. The resulting
object will have unique starts, but ranges (depending on the vector of lengths)
might have overlaps.
}
\examples{
randomGenomicRanges(c(4,89,325), GRobject)


}
